requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.6.2
charset-normalizer==3.4.0
click==8.2.1
colorama==0.4.6
distro==1.9.0
Flask==3.1.1
google-ai-generativelanguage==0.6.4
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.70.0
groq==0.26.0
grpcio==1.72.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==5.3.2
MarkupSafe==3.0.2
mcpi==1.2.1
mcpi-e==0.3.2006.2
openai==1.84.0
pillow==11.1.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pygame==2.5.2
pyjokes==0.7.2
pyparsing==3.2.3
python-docx==1.1.2
python-telegram-bot==21.3
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
termcolor==2.4.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.2.0
urllib3==2.2.3
Werkzeug==3.1.3
